task/task-0/etc/requirements.txt
# Don't change Flask and gunicorn!
Flask
gunicorn
# TODO! Add your 3party package after this line
pillow
numpy